// apps/web/src/api/__generated__/types/GetNotificationsResponseNotificationsNodesItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */
import type { GetNotificationsResponseNotificationsNodesItemContent } from './GetNotificationsResponseNotificationsNodesItemContent';

export type GetNotificationsResponseNotificationsNodesItem = {
  /** Node Cursor */
  cursor: string;
  /** Node Content */
  content: GetNotificationsResponseNotificationsNodesItemContent;
};